Split FAQ question tasks into individual files
