+++ handlers/faq/edit_question_task.go
@@ -1,0 +1,50 @@
+package faq
+
+import (
+	"database/sql"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/handlers"
+	"github.com/arran4/goa4web/internal/db"
+	"github.com/arran4/goa4web/internal/tasks"
+	"github.com/gorilla/mux"
+)
+
+// EditQuestionTask modifies an existing FAQ entry.
+type EditQuestionTask struct{ tasks.TaskString }
+
+var editQuestionTask = &EditQuestionTask{TaskString: TaskEdit}
+var _ tasks.Task = (*EditQuestionTask)(nil)
+
+func (EditQuestionTask) Match(r *http.Request, m *mux.RouteMatch) bool {
+	return tasks.HasTask(TaskEdit)(r, m)
+}
+
+func (EditQuestionTask) Action(w http.ResponseWriter, r *http.Request) any {
+	question := r.PostFormValue("question")
+	answer := r.PostFormValue("answer")
+	category, err := strconv.Atoi(r.PostFormValue("category"))
+	if err != nil {
+		return fmt.Errorf("category parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	faq, err := strconv.Atoi(r.PostFormValue("faq"))
+	if err != nil {
+		return fmt.Errorf("faq id parse fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+
+	if err := queries.UpdateFAQQuestionAnswer(r.Context(), db.UpdateFAQQuestionAnswerParams{
+		Answer:                       sql.NullString{Valid: true, String: answer},
+		Question:                     sql.NullString{Valid: true, String: question},
+		FaqcategoriesIdfaqcategories: int32(category),
+		Idfaq:                        int32(faq),
+	}); err != nil {
+		return fmt.Errorf("update faq question fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+
+	return nil
+}
